Former-commit-id: 3a1a0708897c5d8bbbeb48868fe277415bad8073 [formerly 286b4be9087505bd588f8bfecfd84fd84994d904]
Former-commit-id: 33fc821324adc783b1d48b7d0919a1341cf2d709

+++ public/components/delete-service-dialog/delete-service-dialog.js
@@ -8,14 +8,22 @@ function(Component, initView){
 		tag: 'bh-delete-service-dialog',
 		template: initView,
 		scope : {
+			isDeleting : false,
 			deleteService : function(){
+				if(this.attr('isDeleting')) return;
 				var deleteItems = this.attr('deleteItems') || false;
 				var service = this.attr('service');
+
+				this.attr('isDeleting', true);
 				if(deleteItems){
 					service.destroyIncludingItems();
 				} else {
 					service.destroy();
 				}
+			},
+			cancelDelete : function(){
+				if(this.attr('isDeleting')) return;
+				this.clearService();
 			},
 			clearService : function(){
 				this.attr('service', null);
@@ -24,9 +32,9 @@ function(Component, initView){
 		events : {
 			'{scope.service} destroyed' : function(){
 				if(this.scope.attr('deleteItems')){
-					console.log(this.scope.attr('state'))
 					this.scope.attr('state').resetEmbed();
 				}
+				this.scope.attr('isDeleting', false);
 				this.scope.clearService();
 			}
 		}
